--- bin/application.properties
@@ -1,10 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost/db_metrica_crud?useSSL=false
-spring.datasource.username=root
-spring.datasource.password=metrica
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.database-platform=org.hibernate.dialect.MariaDB103Dialect
-spring.jpa.hibernate.ddl-auto=update
-logging.level.org.hibernate.SQL=debug
-server.port=8081
-
-
