+++ src/main/java/com/metrica/formacion/apigruposmetrica/ApigruposmetricaApplication.java
@@ -1,0 +1,13 @@
+package com.metrica.formacion.apigruposmetrica;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApigruposmetricaApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ApigruposmetricaApplication.class, args);
+    }
+
+}
