+++ src/main/resources/application.properties
@@ -3,19 +3,14 @@ spring.datasource.username=admin
 spring.datasource.password=ARTfer050609
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.database-platform=org.hibernate.dialect.MariaDB103Dialect
+
 spring.jpa.hibernate.ddl-auto=update
-logging.level.org.hibernate.SQL=debug
-#logging.level.org.hibernate.SQL=debug
-server.port=8087
+
+server.port=8083
 hibernate.show_sql=true
-
 spring.main.allow-bean-definition-overriding=true
-
-spring.application.name=api-usuarioss
-
+spring.application.name=api-grupos
 eureka.client.enabled=true
 eureka.client.serviceUrl.defaultZone:http://localhost:1111/eureka
 eureka.instance.leaseRenewalIntervalInSeconds:5
-
-
