--- src/main/resources/import.sql
@@ -1,7 +1,0 @@
-/* Crear Datos para Metricad_db, grupos */
-
-/* Crear Grupos */
-INSERT INTO grupos ( nombre, huecos_libres) VALUES( '2019-01-01 13:00:00', 12);
-INSERT INTO grupos ( nombre, huecos_libres) VALUES( '2019-01-01 14:00:00', 12);
-INSERT INTO grupos ( nombre, huecos_libres) VALUES( '2019-01-01 15:00:00', 12);
-INSERT INTO grupos ( nombre, huecos_libres) VALUES( '2019-01-01 16:00:00', 12);
