+++ src/test/java/com/metrica/formacion/apigruposmetrica/ApigruposmetricaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.metrica.formacion.apigruposmetrica;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApigruposmetricaApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
