--- src/test/java/com/metrica/formacion/dao/GrupoRepositoryTest.java
@@ -1,11 +1,0 @@
-package com.metrica.formacion.dao;
-
-import com.metrica.formacion.SpringConfigurationFile;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest(classes = SpringConfigurationFile.class)
-@Transactional
-class GrupoRepositoryTest {
-
-}
